refactor:updated the folder name
