+++ src/apis/movie/endPoints.ts
